+++ crates/module-system/sov-modules-api/src/runtime/capabilities/kernel.rs
@@ -1,7 +1,6 @@
 use sov_rollup_interface::da::DaSpec;
 use sov_state::Storage;
 
-use super::BlobSelector;
 use crate::{BootstrapWorkingSet, Gas, KernelStateAccessor, Spec, StateCheckpoint};
 
 /// Allows the kernel to map between a slot number and the visible height at that slot.
@@ -38,23 +37,19 @@ pub trait Kernel<S: Spec>: Default {
 }
 
 /// Hooks allowing the kernel to get access to the DA layer state
-pub trait KernelSlotHooks<S: Spec>: BlobSelector<Spec = S> + Kernel<S> {
+pub trait KernelSlotHooks<S: Spec> {
     /// Called at the beginning of a slot. Computes the gas price for the slot
     /// Returns the visible root hash accessible at the current *virtual* rollup height
     fn begin_slot_hook(
         &self,
-        slot_header: &<<Self::Spec as Spec>::Da as DaSpec>::BlockHeader,
-        validity_condition: &<<Self::Spec as Spec>::Da as DaSpec>::ValidityCondition,
-        pre_state_root: &<<Self::Spec as Spec>::Storage as Storage>::Root,
-        state: &mut KernelStateAccessor<'_, <Self::Spec as Spec>::Storage>,
+        slot_header: &<S::Da as DaSpec>::BlockHeader,
+        validity_condition: &<S::Da as DaSpec>::ValidityCondition,
+        pre_state_root: &<S::Storage as Storage>::Root,
+        state: &mut KernelStateAccessor<'_, S::Storage>,
     ) -> <S::Storage as Storage>::Root;
 
     /// Called at the end of a slot
-    fn end_slot_hook(
-        &self,
-        gas_used: &S::Gas,
-        state: &mut KernelStateAccessor<'_, <Self::Spec as Spec>::Storage>,
-    );
+    fn end_slot_hook(&self, gas_used: &S::Gas, state: &mut KernelStateAccessor<'_, S::Storage>);
 
     /// Returns the base fee per gas accessible at the current *virtual* slot.
     fn base_fee_per_gas(&self, state: &mut StateCheckpoint<S::Storage>) -> <S::Gas as Gas>::Price;