+++ examples/demo-rollup/provers/risc0/guest-celestia/src/bin/rollup.rs
@@ -7,7 +7,6 @@ use demo_stf::runtime::Runtime;
 use demo_stf::StfVerifier;
 use sov_celestia_adapter::types::Namespace;
 use sov_celestia_adapter::verifier::{CelestiaSpec, CelestiaVerifier};
-use sov_kernels::basic::BasicKernel;
 use sov_mock_zkvm::{MockZkGuest, MockZkVerifier};
 use sov_modules_api::default_spec::DefaultSpec;
 use sov_modules_api::execution_mode::Zk;
@@ -28,10 +27,9 @@ pub fn main() {
     let stf: StfBlueprint<
         DefaultSpec<CelestiaSpec, Risc0Verifier, MockZkVerifier, Zk>,
         Runtime<_>,
-        BasicKernel<_>,
     > = StfBlueprint::new();
 
-    let stf_verifier = StfVerifier::<_, _, _, _, Risc0Guest, MockZkGuest>::new(
+    let stf_verifier = StfVerifier::<_, _, _, Risc0Guest, MockZkGuest>::new(
         stf,
         CelestiaVerifier {
             rollup_batch_namespace: ROLLUP_BATCH_NAMESPACE,