+++ examples/demo-rollup/stf/src/lib.rs
@@ -13,5 +13,5 @@ use sov_modules_stf_blueprint::StfBlueprint;
 use sov_rollup_interface::stf::StateTransitionVerifier;
 
 /// Alias for StateTransitionVerifier.
-pub type StfVerifier<DA, ZkSpec, RT, K, InnerVm, OuterVm> =
-    StateTransitionVerifier<StfBlueprint<ZkSpec, RT, K>, DA, InnerVm, OuterVm>;
+pub type StfVerifier<DA, ZkSpec, RT, InnerVm, OuterVm> =
+    StateTransitionVerifier<StfBlueprint<ZkSpec, RT>, DA, InnerVm, OuterVm>;